fix: use the right source paths to resolve the right files
ngcc/the Angular CLI follows and compiles the source paths in a fairly
shallow manner - if there's a direct (`./file`) resolution it will
follow that and compile the source. However, if the file path has some
indirection (`@carbon/icons-angular/file`) it _won't_ follow and compile
that path - instead it will resolve it on an as-needed basis as webpack
resolves and compiles the actual source files.

This means that we can have each icon opt-out of the inital ngcc pass
(which saves a ton of time) and reley on the webpack resolution algorithm
to tell ngcc which files to actually compile.

Basically we fake out ngcc by pretending that we want to compile files
from a bunch of different top level packages in node_modules (which
it won't do), but webpack has no qualms about resolving those top level
paths (which are actually just sub-directories) and it will pass everything
to ngcc. TADA! No more ngcc and webpack trying to compile the same modules
twice!